+++ viewplayer.py
@@ -11,7 +11,7 @@ from players import Player
 
 class PlayerPage(webapp2.RequestHandler):
   def get(self, playerID):
-    playerID = int(playerID)
+    playerID = long(playerID)
     p = Player.get_by_id(playerID)
     games = Game.query(Game.players == playerID)
     self.response.write(get_template('viewplayer.html').render({'player': p, 'games': games}))
